Cake price is decimal now
